fix warnings

+++ wp.c
@@ -24,8 +24,8 @@ int main(int argc, char *argv[]) {
     long file_size = ftell(file);
     fseek(file, 0, SEEK_SET);
 
-    char *source = malloc(file_size + 1);
-    fread(source, 1, file_size, file);
+    char *source = malloc((size_t)file_size + 1);
+    fread(source, 1, (size_t)file_size, file);
     source[file_size] = '\0';
     fclose(file);
